requirements.txt
blinker==1.9.0
chardet==5.2.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
et_xmlfile==2.0.0
Flask==3.1.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysql-connector-python==8.0.29
openpyxl==3.1.5
packaging==24.2
pillow==11.2.1
protobuf==6.30.1
pymongo==4.11.2
python-dotenv==1.0.1
reportlab==4.3.1
Werkzeug==3.1.3
